Many updates to fix gcc warnings.
- Removal of type punned casts to resolve strict aliasing problems.

- Initialising more variables to avoid "may be used uninitialised"
  warnings. Note most of these are invalid gcc warnings, but a few
  cases are real when faced with invalid data or abnormal errors.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2348 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1

+++ src/prefinish/finish_pcr.c
@@ -135,7 +135,7 @@ static g4_primer_pair *pick_pcr_primers2(finish_t *fin, primlib_state *pstate,
 		cons_joined[i] = 'N';
     }
     puts(cons_joined);
-    printf("target = %"PRId64",%d\n", strlen(cons1)+1, 20);
+    printf("target = %"PRId64",%d\n", (uint64_t)strlen(cons1)+1, 20);
 
     /* Tweak arguments */
     pstate->p3args.primer_task = pick_pcr_primers;
